Version 2.20
 New Features
 ------------
 - Added new connection parameter called "extra", you can place any
additional command line parameters in this option.
 - DBGetOption displays all options including the dbext version.
 - Better support for command terminators that have newline or special
characters in them.  For example ASE/SQLSRV use "\ngo\n", now the command
DBExecSQLUnderCursor will correctly find the statement.
 - Use_tbl_alias defaults to "ask", instead of "default on".
 - For most supported databases, when displaying the Table / Procedure /
View List, you can now enter a partial string "DBA.", and if a . is included
it will only display objects created / owned by that userid.
 - DBExecSQLUnderCursor would sometimes stop in the middle of a query if the
command terminator was included (inside a quoted string), now it ensures
there is no text following the terminator.
 - The result window also includes the error code returned by the binary
used to execute the command.  This can be useful for debugging.
 - The first line of the result window includes a line showing the
connection information, if you have many buffers open, it can be difficult
to determine which database you are executing commands against.  A glance at
this line will tell you immediately.
 - g:dbext_default_always_prompt_for_variables = 0 will prevent you from
being prompted for substitution parameters.  Setting this value to 1 will
always prompt the user.
 - You can now abort the execution of a statement if you are prompted for
substitution parameters.
 - If you are prompted for parameters, if you choose "Stop Prompting" the
command will be executed immediate.  If "Never Prompt" is chosen, no further
prompting will occur for this buffer.

 Bug Fixes
 ---------
 - SQLSRV did not have a default command terminator, now it is "\ngo\n".
 - Changed the Oracle command terminator to ";", and the routine that
executes the statements automatically adds the "\nquit;\n" so that   sqlplus
will complete.
 - Spaces were not correctly removed from column lists in all cases, this
showed up as an issue with the SQL Intellisense plugin.
 - When executing SELECT statements the INTO clause (if present) is removed
so the results are displayed in the result window.  Refined the removal of
the INTO clause to ensure it does not interfer with an INSERT or MERGE
statement.